prepare cross-platform release
