+++ .sbtopts
@@ -1,0 +1,4 @@
+-J-Xms2048m
+-J-Xmx2048m
+-J-XX:MaxMetaspaceSize=1024m
+-J-XX:MetaspaceSize=760m
