+++ project/build.properties
@@ -1,1 +1,1 @@
-sbt.version = 1.9.8
+sbt.version = 1.9.9
