+++ project/plugins.sbt
@@ -1,11 +1,13 @@
-addSbtPlugin("org.jmotor.sbt" % "sbt-dependency-updates" % "1.2.7")
-
 addSbtPlugin("org.typelevel" % "sbt-tpolecat" % "0.5.0")
 
 addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.5.2")
 
 addSbtPlugin("com.github.sbt" % "sbt-dynver" % "5.0.1")
 
-addSbtPlugin("ch.epfl.scala" % "sbt-scalafix" % "0.11.1")
+addSbtPlugin("ch.epfl.scala" % "sbt-scalafix" % "0.12.0")
 
-addSbtPlugin("org.scoverage" % "sbt-scoverage" % "2.0.9")
+addSbtPlugin("org.scoverage" % "sbt-scoverage" % "2.0.11")
+
+addSbtPlugin("org.portable-scala" % "sbt-scala-native-crossproject" % "1.3.2")
+
+addSbtPlugin("org.scala-native" % "sbt-scala-native" % "0.4.17")
