+++ zio-gc-pubsub/shared/src/main/scala-3/SubscriptionFilter.scala
