+++ zio-gc-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/model/SchemaSettings.scala
