+++ zio-gc-pubsub/shared/src/main/scala/com/anymindgroup/pubsub/serde/Serializer.scala
